File details page and prompt before delete

+++ src/controllers/files-controller.js
@@ -3,6 +3,15 @@ const path = require("path");
 const fs = require("fs/promises");
 
 /** @type {import("express").RequestHandler} */
+const getFileDetails = (req, res, next) => {
+  res.render("layout", {
+    template: "pages/file-details",
+    title: "Detalhes do Arquivo",
+    file: res.locals.file,
+  });
+};
+
+/** @type {import("express").RequestHandler} */
 const uploadFile = async (req, res, next) => {
   try {
     const reloadDirectory = () =>
@@ -41,9 +50,26 @@ const downloadFile = async (req, res, next) => {
 };
 
 /** @type {import("express").RequestHandler} */
-const deleteFile = async (req, res, next) => {
+const promptDeleteFile = (req, res, next) => {
+  res.render("layout", {
+    template: "pages/prompt",
+    title: "Apagar Arquivo",
+    promptTitle: "Apagar Arquivo",
+    promptText: `Tem certeza que deseja apagar o arquivo "${res.locals.file.fileName}"?`,
+    action: `/arquivo/${res.locals.file.id}/apagar`,
+  });
+};
+
+/** @type {import("express").RequestHandler} */
+const postDeleteFile = async (req, res, next) => {
   try {
     const { id, nameOnStorage, directoryId } = res.locals.file;
+    const deleteConfirmed = req.body.response === "true";
+
+    if (!deleteConfirmed) {
+      return res.redirect(`/pasta/${directoryId}`);
+    }
+
     await fs.unlink(
       path.join(__dirname, `../../multer-uploads/${nameOnStorage}`)
     );
@@ -55,9 +81,11 @@ const deleteFile = async (req, res, next) => {
 };
 
 const filesController = {
+  getFileDetails,
   uploadFile,
   downloadFile,
-  deleteFile,
+  promptDeleteFile,
+  postDeleteFile,
 };
 
 module.exports = filesController;
